BioIDMapper-package {BioIDMapper}	R Documentation

Mapping biological ids and linking to external data source

Description

This package will map ids among different biological ids from NCBI and Uniprot. Ids will be from gene level, protein level, functional level and metabolomic level.

Details

Package: 	 BioIDMapper

Type: 	 Package

Version: 	 1.0

Date: 	 2007-09-05

License: 	 What license is it under?

The main interface is bio.convert() function

Author(s)

Xiaoyong Sun Maintainer: johnsunx1<johnsunx1@gmail.com>

Examples


data(glist)
bio.convert(glist, 1, 6)

[Package BioIDMapper version 1.0 Index]