bio.convert {BioIDMapper}	R Documentation

Main Interface for Mapping

Description

This function is main interface for mapping service. It can support NCBI mapping and UniProt mapping, from gene level, protein level, functional level, and metabolomic level.

Usage


bio.convert(id_list = "character", from = "numeric", to = "numeric")

Arguments

id_list 	 the list of biological ids

from 	 Biokey number for source biological type. It can be found by bio.type()

to 	 Biokey number for destination biological type. It can be found by bio.type()

Details

It converts a list of ids from one biological type to another. For detail, please check vegnette "demonstration" section.

Value

a result matrix will be returned with source type in the first column and the destination type in the last column

Author(s)

Xiaoyong Sun

Examples



data(glist)
bio.type(1)
bio.type(5)
bio.convert(glist, 1, 5)->done

[Package BioIDMapper version 1.0 Index]