bio.select {BioIDMapper}	R Documentation

select subset from mapped matrix

Description

use id as paramter, this function can select subste from mapping result

Usage


bio.select(result_matrix, colno, myid)

Arguments

result_matrix 	 the returned matrix from bio.convert()

colno 	 the column number from returned matrix, which contains the id you are interesed

myid 	 id you are interesed

Author(s)

Xiaoyong Sun

Examples


data(glist)
bio.convert(glist, 1, 5)->done
bio.select(done, 1, "200529" )

[Package BioIDMapper version 1.0 Index]