bio.type {BioIDMapper}	R Documentation

Get the biokey number this package support

Description

show the biokey number this package support

Usage


bio.type(type_to_id)

Arguments

type_to_id 	 If no parameter is used, return tables for all biokey numbers; if parameter is number from bio.type(), return the corresponding biological type; if parameter is biological type, return the corresponding biokey number

Author(s)

Xiaoyong Sun

Examples



bio.type()
bio.type(5)
bio.type("PDB id")


[Package BioIDMapper version 1.0 Index]